change to class
