+++ blog/src/Comments.js
@@ -1,0 +1,25 @@
+import React from "react";
+import "./App.css"
+
+class Comments extends React.Component{
+    constructor(props) {
+        super(props);
+        this.state={result:[]};
+    }
+
+    
+
+    deleteComment(){
+
+    }
+
+    render() {
+        return(
+          <div>
+
+          </div>
+        );
+    }
+}
+
+export default Comments;
