+++ blog/src/Memo.js
@@ -1,12 +1,75 @@
 import React, { useState } from "react";
 import "./App.css";
 import Loading from "./Loading";
-import {RefreshComments} from "./RefreshComments";
 
 const Memo = (props) => {
   // const [address, setAddress] = useState("///당신의 위치를 고르세요!");
+    let wids = [];
+    let wnames = [];
+    let wpasses = [];
+    let wdates = [];
+    let contents = [];
 
-  // 이게 기본값임.
+    const RefreshComments = () => {
+        let resultArr = [];
+        while(resultArr.length>0){
+            resultArr.pop();
+        }
+        fetch('/showComments.php?w3=' + props.address)
+            .then(function (response) {
+                return response.json();
+            })
+            .then(function (json) {
+                for (let i = 0; i < json.length; i++) {
+                    wids.push(json[i].wid);
+                    wnames.push(json[i].wname);
+                    wpasses.push(json[i].wpass);
+                    wdates.push(json[i].wdate);
+                    contents.push(json[i].content);
+                }
+            })
+        for(let i=0;i<wids.length;i++){
+            resultArr.push(
+                <table border={3} width={600} align={"center"}>
+                    <tr>
+                        <td>
+                            wid : {wids[i]}
+                        </td>
+                        <td>
+                            wname : {wnames[i]}
+                        </td>
+                        <td>
+                            wpass : {wpasses[i]}
+                        </td>
+                        <td>
+                            wdate : {wdates[i]}
+                        </td>
+                        <td>
+                            content : {contents[i]}
+                        </td>
+                        <td>
+                            <button onClick={()=>deleteComment(wids[i], wpasses[i])}>
+                                삭제
+                            </button>
+                        </td>
+                    </tr>
+                </table>);
+        }
+        console.log(resultArr);
+        return (resultArr);
+    }
+
+    const deleteComment = (wid, wpass) => {
+        let inputPass = prompt("비밀번호를 입력하세요");
+        if (inputPass === wpass) {
+            let response = fetch("/deleteComment.php?wid=" + wid);
+            alert("글이 삭제되었습니다!");
+        } else {
+            alert("비밀번호가 틀렸습니다!");
+        }
+    }
+
+    // 이게 기본값임.
   return (
     <div className="ContentBox" style={{ backgroundColor: props.color }}>
       <h2>Content</h2>
@@ -51,7 +114,7 @@ const Memo = (props) => {
           </tr>
           <tr>
             <td colSpan={4} align={"right"}>
-              <input type={"submit"} value={"+"} />
+              <input type={"submit"} value={"+"} onClick={()=>RefreshComments()}/>
             </td>
           </tr>
         </table>
@@ -62,10 +125,12 @@ const Memo = (props) => {
       {/* 지도의 '변환하기' 버튼을 클릭하면 작동 / 일단은 임의의 버튼으로 구현 */}
       <h2>방문글 목록</h2>
         <RefreshComments address={props.address} />
+        <button onClick={() => RefreshComments()}>refresh</button>
       <Loading />
     </div>
   );
 };
+
 
 Memo.defaultProps = {
   color: "midnightblue",
